+++ src/components/Screen/components/index.ts
@@ -1,0 +1,2 @@
+export * from './ScreenContainer';
+export * from './ScreenHeader';
